requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
Django==4.2.13
django-bootstrap4==22.2
django-crispy-forms==2.1
pillow==10.3.0
psycopg==3.1.19
soupsieve==2.5
sqlparse==0.5.0
typing-extensions==4.11.0
